backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.40.0
accelerate==0.25.0
einops==0.7.0
pydantic==2.5.2
aiofiles==23.2.1
python-dotenv==1.0.0
mss==9.0.1
pyautogui==0.9.54
selenium==4.16.0
webdriver-manager==4.0.1
numpy==1.24.3
opencv-python==4.8.1.78
sentencepiece>=0.1.99
protobuf>=3.20.0
timm>=1.0.0